backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
email-validator==2.2.0
pytest==8.3.2
httpx==0.27.2
